requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.1
Django==4.2.20
django-allauth==65.7.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
exceptiongroup==1.2.2
executing==2.2.0
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
ipython==8.18.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.8
matplotlib-inline==0.1.7
mysqlclient==2.2.7
parso==0.8.4
prompt_toolkit==3.0.51
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
zipp==3.21.0
